Index

GasChem.FastJX

GasChem.FastJX_interpolation_troposphere

GasChem.GEOSChemGasPhase

GasChem.SuperFast

GasChem.adjust_j_o31D

GasChem.arr_3rd

GasChem.arr_3rdbody

GasChem.arrh

GasChem.arrhenius

GasChem.arrplus

GasChem.constant_k

GasChem.cos_solar_zenith_angle

GasChem.create_fjx_interp

GasChem.eq_const

GasChem.j_mean

GasChem.rate_2HO2

GasChem.rate_ALK

GasChem.rate_DMSOH

GasChem.rate_EPO

GasChem.rate_GLYCOH_a

GasChem.rate_GLYCOH_b

GasChem.rate_GLYXNO3

GasChem.rate_HACOH_a

GasChem.rate_HACOH_b

GasChem.rate_HO2HO2

GasChem.rate_ISO1

GasChem.rate_ISO2

GasChem.rate_NIT

GasChem.rate_OHCO

GasChem.rate_OHHNO3

GasChem.rate_OH_CO

GasChem.rate_RO2HO2

GasChem.rate_RO2NO_a1

GasChem.rate_RO2NO_a2

GasChem.rate_RO2NO_b1

GasChem.rate_RO2NO_b2

GasChem.rate_toppb

GasChem.tbranch

GasChem.tunplus

API Documentation

GasChem.FastJX — Method

Description: This is a box model used to calculate the photolysis reaction rate constant using the Fast-JX scheme (Neu, J. L., Prather, M. J., and Penner, J. E. (2007), Global atmospheric chemistry: Integrating over fractional cloud cover, J. Geophys. Res., 112, D11306, doi:10.1029/2006JD008007.)

Argument:

t_ref: Reference time for the model, can be a DateTime or a Unix timestamp (in seconds).

Example

Build Fast-JX model:

fj = FastJX(DateTime(2000, 1, 1))

GasChem.FastJX_interpolation_troposphere — Method

Description: This is a box model used to calculate the photolysis reaction rate constant using the Fast-JX scheme (Neu, J. L., Prather, M. J., and Penner, J. E. (2007), Global atmospheric chemistry: Integrating over fractional cloud cover, J. Geophys. Res., 112, D11306, doi:10.1029/2006JD008007.)

Argument:

t_ref: Reference time for the model, can be a DateTime or a Unix timestamp (in seconds).

Example

Build Fast-JX model:

fj = FastJX(DateTime(2000, 1, 1))

GasChem.GEOSChemGasPhase — Method

GEOS-Chem full-chem mechanism adapted from GEOS-Chem version 14.1.1

Adapted from file https://github.com/geoschem/geos-chem/blob/4722f288e90291ba904222f4bbe4fc216d17c34a/KPP/fullchem/fullchem.eqn

The GEOS-Chem license applies: https://github.com/geoschem/geos-chem/blob/main/LICENSE.txt

If the keyword argument rxn_sys is set to true, the function will return a reaction system instead of an ODE system.

===============================================================================

REFERENCES (alphabetical order)

Atkinson2003: Atkinson and Arey, Chem. Rev., doi:10.1021/cr0206420, 2003.

Bates2014: Bates et al., J. Phys. Chem A, 118, doi:10.1021/jp4107958, 2014.

Bates2019: Bates and Jacob, Atmos. Chem. Phys., doi:10.5194/acp-19-9613-2019, 2019.

Bates2021a: Bates et al, JGR, https://doi.org/10.1029/2020JD033439, 2021.

Bates2021b: Bates et al, ACP, https://doi.org/10.5194/acp-2021-605, 2021.

Browne2011: Browne et al., Atmos. Chem. Phys., doi:10.5194/acp-11-4209-2011, 2011.

Browne2014: Browne et al., Atmos. Chem. Phys., doi:10.5194/acp-14-1225-2014, 2014.

Chen2017: Chen et al., Geophys. Res. Lett., doi:10.1002/2017GL073812, 2017.

Crounse2012: Crounse et al., J. Phys. Chem. A, doi:10.1021/jp211560u, 2012.

Eastham2014: Eastham et al., Atmos. Env., doi:10.1016/j.atmosenv.2014.02.001, 2014.

Fischer2014: Fischer et al., Atmos. Chem. Phys., doi:10.5194/acp-14-2679-2014, 2014.

Fisher2016: Fisher et al., Atmos. Chem. Phys., doi:10.5194/acp-16-5969-2016, 2016.

Fisher2018: Fisher et al., J. Geophys. Res., doi:10.1029/2018JD029046, 2018.

Fry2014: Fry et al. Environ. Sci. Technol., doi:10.1021/es502204x, 2014.

Gill2002: Gill and Hites, J. Phys. Chem. A, doi:10.1021/jp013532, 2002.

Goliff2013: Goliff et al., Atmos. Environ., doi:10.1016/j.atmosenv.2012.11.038, 2013.

Jacobs2014: Jacobs et al., Atmos. Chem. Phys., doi:10.5194/acp-14-8933-2014, 2014.

Jenkin2015: Jenkin et al., Atmos. Chem. Phys., doi:10.5194/acp-15-11433-2015, 2015.

Kasibhatla2018: Kasibhatla et al., Atmos. Chem. Phys., doi:10.5194/acp-18-11185-2018, 2018

IUPAC ROO19: https://iupac-aeris.ipsl.fr/htdocs/datasheets/pdf/ROO19CH3O2NO3.pdf

JPL 10-6: JPL Publication 10-6, https://jpldataeval.jpl.nasa.gov/previous_evaluations.html, 2011.

JPL 15-10: JPL Publication 15-10, https://jpldataeval.jpl.nasa.gov, 2015.

Kwon2020: Kwon et al, Elementa, https://doi.org/10.1525/elementa.2021.00109, 2020.

Lee2014: Lee et al., J. Phys. Chem. A, doi:10.1021/jp4107603, 2014.

Marais2016: Marais et al., Atmos. Chem. Phys, doi:10.5194/acp-16-1603-2016, 2016.

Miller2017: Miller et al., Atmos. Chem. Phys. Discuss., doi:10.5194/acp-2016-1042, 2017.

Millet2015: Millet et al., Atmos. Chem. Phys., doi:10.5194/acp-15-6283-2015, 2015. Moch et al, JGR, https, * Moch2020 # //doi.org/10.1029/2020JD032706, 2020.

Muller2014: Muller et al., Atmos. Chem. Phys., doi:10.5194/acp-14-2497-2014, 2014.

Parrella2012: Parrella et al. Atmos. Chem. Phys, doi:10.5194/acp-12-6723-2012, 2012.

Paulot2009: Paulot et al., Atmos. Chem. Phys., doi:10.5194/acp-9-1479-2009, 2009a and Paulot et al., Science, doi:10.1126/science.1172910, 2009b.

Peeters2010: Peeters and Muller, Phys. Chem. Chem. Phys., doi:10.1039/C0CP00811G, 2010.

Peeters2014: Peeters et al., J. Phys. Chem. A, doi:10.1021/jp5033146, 2014.

Pye2010: Pye et al., Atmos. Chem. Phys., doi:10.5194/acp-10-11261-2010, 2010.

Roberts1992: Roberts and Bertman, Int. J. Chem. Kinet., doi:10.1002/kin.550240307, 1992.

Sherwen2016b: Sherwen et al., Atmos. Chem. Phys., doi:10.5194/acp-16-12239-2016, 2016b.

Sherwen2017: Sherwen et al., Faraday Discuss., doi:10.1039/C7FD00026J, 2017.

StClair2016: St. Clair et al., J. Phys. Chem. A, doi:10.1021/acs.jpca.5b065322016, 2016.

Travis2016: Travis et al., Atmos. Chem. Phys., doi:10.5194/acp-16-13561-2016, 2016.

Wolfe2012: Wolfe et al., Phys. Chem. Chem. Phys., doi: 10.1039/C2CP40388A, 2012.

Xie2013: Xie et al., Atmos. Chem. Phys., doi:10.5194/acp-13-8439-2013, 2013.

GasChem.SuperFast — Method

SuperFast()

This atmospheric chemical system model is built based on the Super Fast Chemical Mechanism, which is one of the simplest representations of atmospheric chemistry. It can efficiently simulate background tropheric ozone chemistry and perform well for those species included in the mechanism. The chemical equations we used is included in the supporting table S2 of the paper:

"Evaluating simplified chemical mechanisms within present-day simulations of the Community Earth System Model version 1.2 with CAM4 (CESM1.2 CAM-chem): MOZART-4 vs. Reduced Hydrocarbon vs. Super-Fast chemistry" (2018), Benjamin Brown-Steiner, Noelle E. Selin, Ronald G. Prinn, Simone Tilmes, Louisa Emmons, Jean-François Lamarque, and Philip Cameron-Smith.

The input of the function is Temperature, concentrations of all chemicals, and reaction rates of photolysis reactions

If the keyword argument rxn_sys is set to true, the function will return a reaction system instead of an ODE system.

Example

using GasChem, EarthSciMLBase, DifferentialEquations, Plots
rs = SuperFast()
sol = solve(ODEProblem(structural_simplify(rs), [], (0,360), [], combinatoric_ratelaws=false), AutoTsit5(Rosenbrock23()), saveat=10.0)
plot(sol)

GasChem.adjust_j_o31D — Method

adjust_j_o31D(T, P, H2O)

Adjust the photolysis rate of O3 -> O2 + O(1D) to represent the effective rate for O3 -> 2OH. This adjustment is based on the fraction of O(1D) that reacts with H2O to produce 2 OH.

GasChem.arr_3rd — Method

Third body effect for pressure dependence of rate coefficients. a1, b1, c1 are the Arrhenius parameters for the lower-limit rate. a2, b2, c2 are the Arrhenius parameters for the upper-limit rate. fv is the falloff curve paramter, (see ATKINSON ET. AL (1992) J. Phys. Chem. Ref. Data 21, P. 1145). Usually fv = 0.6.

GasChem.arr_3rdbody — Method

Third body effect for pressure dependence of rate coefficients. a1, b1, c1 are the Arrhenius parameters for the lower-limit rate. a2, b2, c2 are the Arrhenius parameters for the upper-limit rate. fv is the falloff curve paramter, (see ATKINSON ET. AL (1992) J. Phys. Chem. Ref. Data 21, P. 1145). Usually fv = 0.6.

GasChem.arrh — Method

Arrhenius equation:

\[ k = a0 * exp( c0 / T ) * (T/300)^b0\]

GasChem.arrhenius — Method

Arrhenius equation:

\[ k = a0 * exp( c0 / T ) * (T/300)^b0\]

GasChem.arrplus — Method

Modified Arrhenius law.

GasChem.constant_k — Method

Function to create a constant rate coefficient

GasChem.cos_solar_zenith_angle — Method

cos_solar_zenith_angle(lat, t, long)

This function is to compute the cosine of the solar zenith angle, given the unixtime, latitude and longitude The input variables: lat=latitude(°), long=longitude(°), t=unixtime(s) the cosine of the solar zenith angle (SZA) is given by: . cos(SZA) = sin(LAT)sin(DEC) + cos(LAT)cos(DEC)*cos(AHR)

       where LAT = the latitude angle,
             DEC = the solar declination angle,
             AHR = the hour angle, all in radians.  All in radians

GasChem.create_fjx_interp — Method

Create a vector of interpolators to interpolate the cross sections σ (TODO: What are the units?) for different wavelengths (in nm) and temperatures (in K).

We use use linear interpolation with flat extrapolation.

GasChem.eq_const — Method

Calculates the equilibrium constant Find the backwards reaction by K=kforward/kbackwards Calculates the rate constant of the forward reaction

Used to compute the rate for these reactions: PPN = RCO3 + NO2 PAN = MCO3 + NO2 ClOO {+M} = Cl + O2 {+M} Cl2O2 {+M} = 2ClO {+M}

GasChem.j_mean — Method

Get mean photolysis rates at different times

GasChem.rate_2HO2 — Method

Used to compute the rate for this reactions: HO2 + HO2 = H2O2 + O2

GasChem.rate_ALK — Method

Used to compute the rate for these reactions: IHOO1 + NO = NO2 + ... IHOO4 + NO = NO2 + ... IHP001 + NO = NO2 + ... IHP002 + NO = NO2 + ... IHP003 + NO = NO2 + ... IEPOXAOO + NO = NO2 + ... IEPOXBOO + NO = NO2 + ... ICHOO + NO = NO2 + ... ISOPNOO1 + NO = 1.728NO2 + ... ISOPNOO2 + NO = NO2 + ... IDHNDOO1 + NO = NO2 + ... IDHNDOO2 + NO = NO2 + ... IDHNBOO + NO = NO2 + ... IDHNDOO + NO = NO2 + ... INO2B + NO = 2.000NO2 + ... INO2D + NO = NO2 + ... IHPNBOO + NO = 1.065NO2 + ... IHPNDOO + NO = NO2 + ... MVKOHOO + NO = NO2 + ... MCROHOO + NO = NO2 + ...

GasChem.rate_DMSOH — Method

Reaction rate for: DMS + OH = 0.750SO2 + 0.250MSA + MO2

GasChem.rate_EPO — Method

Used to compute the rate for these reactions: RIPA + OH = 0.67IEPOXA + 0.33IEPOXB + OH + 0.005LVOC RIPB + OH = 0.68IEPOXA + 0.321IEPOB + OH + 0.005LVOC IEPOXA + OH = 0.67IEPOXA00 + 0.33IEPOXB00 IEPOXB + OH = 0.81IEPOXA00 + 0.19IEPOXB00 IHN2 + OH = 0.67IEPOXA + 0.33IEPOXB + NO2 IHN3 + OH = 0.67IEPOXA + 0.33IEPOXB + NO2 IHN1 + OH = IEPOXD + NO2 IHN4 + OH = IEPOXD + NO2 INPB + OH = OH + ITHN INPD + OH = OH + ITHN INPD + OH = NO2 + ICHE ICN + OH = NO2 + ICHE

GasChem.rate_GLYCOH_a — Method

Used to compute the rate for this reaction: GLYC + OH = 0.732CH2O + 0.361CO2 + 0.505CO + 0.227OH

0.773HO2 + 0.134GLYX + 0.134HCOOH which is the "A" branch of GLYC + OH.

For this reaction, these Arrhenius law terms evaluate to 1: (300/T)^b0 * exp(c0/T) Because b0 = c0 = 0.

GasChem.rate_GLYCOH_b — Method

Used to compute the rate for this reaction: GLYC + OH = HCOOH + OH + CO which is the "B" branch of GLYC + OH.

For this reaction, these Arrhenius law terms evaluate to 1: (300/T)^b0 * exp(c0/T) Because b0 = c0 = 0.

GasChem.rate_GLYXNO3 — Method

Reaction rate for: GLYX + NO3 = HNO3 + HO2 + 2CO i.e. the HO2 + 2*CO branch

GasChem.rate_HACOH_a — Method

Used to compute the rate for this reaction: HAC + OH = MGLY + HO2 which is the "A" branch of HAC + OH.

GasChem.rate_HACOH_b — Method

Used to compute the rate for this reaction: HAC + OH = 0.5HCOOH + OH + 0.5ACTA + 0.5CO2 + 0.5CO + 0.5MO2 which is the "B" branch of HAC + OH.

GasChem.rate_HO2HO2 — Method

Used to compute the rate for this reactions: HO2 + HO2 = H2O2 + O2

GasChem.rate_ISO1 — Method

Used to compute the rate for these reactions: ISOP + OH = LISOPOH + IHOO1 ISOP + OH = LISOPOH + IHOO4

GasChem.rate_ISO2 — Method

Used to compute the rate for these reactions: ISOP + OH = 0.3MCO3 + 0.3MGLY + 0.3CH2O

0.15HPALD3 + 0.25HPALD1 + 0.4HO2

0.6CO + 1.5OH + 0.3HPETHNL + LISOPOH ISOP + OH = 0.3CH2O + 0.15HPALD4 + 0.25HPALD2

1.5OH + 0.9CO + 0.7HO2 + 0.3MGLY

0.3ATOOH + LISOPOH

GasChem.rate_NIT — Method

Used to compute the rate for these reactions: IHOO1 + NO = IHN2 IHOO4 + NO = IHN4 IHPOO1 + NO = IHTN IHPOO2 + NO = IHTN IHPOO2 + NO = IHTN IEPOXAOO + NO = IHTN IEPOXBOO + NO = IHTN IHCOO + NO = IHTN ISOPNOO1 + NO = IDN ISOPNOO2 + NO = IDN IDHNDOO1 + NO = IDN IDHNDOO2 + NO = IDN INO2B + NO = IDN INO2D + NO = IDN IHPNBOO + NO = IDN IHPNDOO + NO = IDN MVK0HOO + NO = 0.438MVKN MCROHOO + NO = MCRHN

GasChem.rate_OHCO — Method

Reaction rate for: OH + CO = HO2 + CO2 (cf. JPL 15-10)

GasChem.rate_OHHNO3 — Method

Used to compute the rate for these reactions: HNO3 + OH = H2O + NO3 HONIT + OH = NO3 + HAC

GasChem.rate_OH_CO — Method

Reaction rate for: OH + CO = HO2 + CO2 (cf. JPL 15-10)

GasChem.rate_RO2HO2 — Method

Carbon Dependence of RO2+HO2, used in these reactions: A3O2 + HO2 = RA3P PO2 + HO2 = PP KO2 + HO2 = 0.150OH + 0.150ALD2 + 0.150MCO3 + 0.850ATOOH B3O2 + HO2 = RB3P PRN1 + HO2 = PRPN

GasChem.rate_RO2NO_a1 — Method

Reaction rate for the "A" branch of these RO2 + NO reactions: MO2 + NO = MENO3 in which the "a1" parameter equals exactly 1.

For these reactions, these Arrhenius law terms evaluate to 1: (300/T)^b0 (300/T)^b1 * exp(c1/T) because b0 = b1 = c1 = 0.

Special treatment for methyl nitrate based on observations as Carter and Atkinson formulation does not apply to C1. Value based on upper limit of Flocke et al. 1998 as applied in Fisher et al. 2018

GasChem.rate_RO2NO_a2 — Method

Reaction rate for the "A" branch of these RO2 + NO reactions, ETO2 + NO = ETNO3 A3O2 + NO = NPRNO3 R4O2 + NO = R4N2 B3O2 + NO = IPRNO3 in which the "a1" parameter is greater than 1.0.

GasChem.rate_RO2NO_b1 — Method

Reaction rate for the "B" branch of these RO2 + NO reactions: MO2 + NO = CH2O + NO2 + HO2 in which the "a1" parameter equals exactly 1.

For these reactions, these Arrhenius law terms evaluate to 1: (300/T)^b0 (300/T)^b1 * exp(c1/T) because b0 = c0 = c1 = 0.

GasChem.rate_RO2NO_b2 — Method

Reaction rate for the "B" branch of these RO2 + NO reactions: ETO2 + NO = NO2 + HO2 + ... A3O2 + NO = NO2 + HO2 + ... R4O2 + NO = NO2 + 0.27HO2 + ... B3O2 + NO = NO2 + HO2 + ... in which the "a1" parameter is greater than 1.0.

Use this function when a1 input argument is greater than 1.0.

GasChem.rate_toppb — Method

Convert reaction rate value in unit of cm^3/molec/s to (s*ppb)^-1

GasChem.tbranch — Method

Temperature Dependent Branching Ratio

GasChem.tunplus — Method

Used to compute the rate for these reactions: IHOO1 = 1.5OH + ... IHOO4 = 1.5OH + ...